fix(build): macOS narrowing error + CI parity configure; guard narrowing
CI (the macOS job) caught a real portability bug: `MethodLocator{GetDexId(),
...}` brace-initializes a uint16_t `dex_id` field from GetDexId()'s uint32_t —
ill-formed C++ that Apple Clang rejects (`-Wc++11-narrowing`) while GCC only
warns, so the Linux dev build silently accepted it. dexllm did not actually
build on macOS.

- dexitem_code_source.cpp: static_cast<uint16_t>(GetDexId()) at both
  MethodLocator init sites (LocateMethod, LocateClassMethods).
- CMakeLists: add -Werror=narrowing to our targets (dexkit_ext / dexkit_dad /
  _dexkit_core; not the vendored Core) so the Linux build now rejects the same
  narrowing macOS would, instead of deferring the discovery to CI.
- ci.yml: build the parity tests in a dedicated `build-ci` CMake tree with an
  explicit pybind11_DIR. Re-driving the editable-install build dir failed to
  re-configure because its pybind11 lived in scikit-build-core's (now-gone)
  isolated build env.

Verified: full build clean under -Werror=narrowing (binding + parity), parity
25/25 via both the install dir and the exact CI `build-ci` command, import OK.

+++ native/core_ext/dexitem_code_source.cpp
@@ -378,7 +378,7 @@ DexItemCodeSource::LocateMethod(std::string_view descriptor) {
         if (strings[m.name_idx] != name) continue;
         if (BuildProto(*dex_item, proto_ids[m.proto_idx]) != proto) continue;
         return dexkit::dad::IDexCodeSource::MethodLocator{
-            dex_item->GetDexId(), static_cast<uint32_t>(i)};
+            static_cast<uint16_t>(dex_item->GetDexId()), static_cast<uint32_t>(i)};
     }
     return std::nullopt;
 }
@@ -391,7 +391,7 @@ DexItemCodeSource::LocateClassMethods(std::string_view class_descriptor) {
     const auto& class_method_ids = dex_item->GetClassMethodIds(type_idx);
     out.reserve(class_method_ids.size());
     for (uint32_t midx : class_method_ids) {
-        out.push_back({dex_item->GetDexId(), midx});
+        out.push_back({static_cast<uint16_t>(dex_item->GetDexId()), midx});
     }
     return out;
 }
